+++ src/store/CHANGELOG.md
@@ -5,10 +5,10 @@ CHANGELOG
 ---
 
  * [BC BREAK] Change `public array $calls` to `private array $calls` in `TraceableStore` - use `getCalls()` instead
+ * Add `count` method to `StoreInterface`
 
 0.7
 ---
-
  * Add `TraceableStore` profiler decorator moved from AI Bundle
  * Add `RstLoader` and `RstToctreeLoader` for loading RST files and following toctree directives
  * Add pre-query event dispatching for query enhancement before vectorization