+++ src/store/src/Bridge/ChromaDb/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Introduce a `StoreFactory`
+ * Add `count` method
 
 0.4
 ---